feat: reorganize classes
move classes into separate files
